backend/requirements.txt
aiogram==3.3.0
SQLAlchemy==2.0.27
aiosqlite==0.19.0
python-dotenv==1.0.1
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
aiohttp==3.9.3
Pillow==10.0.0
pydantic==2.5.3
httpx==0.26.0 